fix: now when unauthenticated user is trying to view, create, update or delete notes or tags 401 error is appropriately raised (used to be 500)

+++ backend/api/views/notes.py
@@ -30,7 +30,7 @@ def try_except(view):
         except AssertionError as e:
             print(e)
             return Response({'errors': e.args[0], 'detail': e.args[0]}, status=status.HTTP_400_BAD_REQUEST)
-    
+        
         except Exception as e:
             print(e)
             return Response({ 'errors': e.args[0], 'detail': e.args[0] }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
@@ -44,6 +44,9 @@ class NoteList(APIView):
     @try_except
     def get(self, request: HttpRequest):
         user = request.user
+
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to view notes')
 
         # params
         page_string = request.GET.get('page')
@@ -80,6 +83,9 @@ class NoteList(APIView):
     @try_except
     def post(self, request: HttpRequest):
         user = request.user
+
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to create notes')
         
         # create new note
         note_form = NoteForm(request.data)
@@ -110,6 +116,9 @@ class NoteDetail(APIView):
     @try_except
     def get(self, request: HttpRequest, pk: str):
         user = request.user
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to view notes')
+
         note = user.note_set.get(id=pk)
         serializer = NoteSerializer(note, many=False)
         return Response(serializer.data)    
@@ -118,6 +127,9 @@ class NoteDetail(APIView):
     @try_except
     def put(self, request: HttpRequest, pk: str):
         user = request.user
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to update notes')
+
         note = user.note_set.get(id=pk)
         note_orig = user.note_set.get(id=pk)
 
@@ -147,6 +159,9 @@ class NoteDetail(APIView):
     @try_except
     def delete(self, request: HttpRequest, pk: str,):
         user = request.user
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to delete notes')
+
         note = user.note_set.get(id=pk)
         note.delete()
