+++ backend/api/views/tags.py
@@ -21,11 +21,11 @@ def try_except(view):
         except Tag.DoesNotExist as e:
             print(e)
             return Response({ 'errors': e.args[0] }, status=status.HTTP_404_NOT_FOUND)
-    
+
         except exceptions.APIException as e:
             print(e)
             return Response({ 'errors': e.detail }, status=e.status_code)
-    
+
         except Exception as e:
             print(e)
             return Response({ 'errors': e.args[0] }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
@@ -39,6 +39,9 @@ class TagList(APIView):
     @try_except
     def get(self, request: HttpRequest):
         user = request.user
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to view tags')
+
         tags = user.tag_set.all()
         serializer = TagSerializer(tags, many=True)
     
@@ -47,7 +50,10 @@ class TagList(APIView):
     @permission_classes([IsAuthenticated])
     @try_except
     def post(self, request: HttpRequest):
-        user = request.user        
+        user = request.user
+
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to create tags')
        
         # create new tag
         tag_form = TagForm(request.data)
@@ -71,6 +77,9 @@ class TagDetail(APIView):
     @try_except
     def get(self, request: HttpRequest, pk: str):
         user = request.user
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to view tags')
+
         tag = user.tag_set.get(id=pk)
         serializer = TagSerializer(tag, many=False)
         return Response(serializer.data)    
@@ -79,6 +88,9 @@ class TagDetail(APIView):
     @try_except
     def put(self, request: HttpRequest, pk: str):
         user = request.user
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to update tags')
+
         tag = user.tag_set.get(id=pk)
         tag_orig = user.tag_set.get(id=pk)
 
@@ -103,6 +115,9 @@ class TagDetail(APIView):
     @try_except
     def delete(self, request: HttpRequest, pk: str,):
         user = request.user
+        if not user or not user.is_authenticated:
+            raise exceptions.NotAuthenticated('You must be logged in to delete tags')
+
         tag = user.tag_set.get(id=pk)
         tag.delete()
